Conflicts:
	pellets/z_win_utils/cmdline.hpp

+++ pellets/z_win_utils/cmdline.hpp
@@ -87,7 +87,7 @@ namespace WinUtils
 
                 }else if(sVal[0] == _T(' ') || _tcslen(sVal) == 1 ) 
                 { // Key with no value or cmdline ends with /Key:
-                    CString csKey(sArg, sVal - sArg);
+                    CString csKey(sArg, (int)(sVal - sArg));
                     if(!csKey.IsEmpty()) 
                     { // Prevent /: case
                         if(!m_bCaseSensitive) 
@@ -102,7 +102,7 @@ namespace WinUtils
                     continue;
                 }else 
                 { // Key with value
-                    CString csKey(sArg, sVal - sArg);
+                    CString csKey(sArg, (int)(sVal - sArg));
                     if(!m_bCaseSensitive) 
                     {
                         csKey.MakeLower();
@@ -133,7 +133,7 @@ namespace WinUtils
                     { // End quote or space present
                         if(!csKey.IsEmpty()) 
                         {	// Prevent /:"val" case
-                            CString csVal(sQuote, sEndQuote - sQuote);
+                            CString csVal(sQuote, (int)(sEndQuote - sQuote));
                             m_ValsMap.insert(CValsMap::value_type(csKey, csVal));
                         }
